Add documentation to the http middleware base class. Reorg files.
