+++ Prometheus.HttpExporter.AspNetCore/HttpRequestCount/HttpRequestCountMiddleware.cs
@@ -1,13 +1,14 @@
 using System;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
+using Prometheus.HttpExporter.AspNetCore.Library;
 
 namespace Prometheus.HttpExporter.AspNetCore.HttpRequestCount
 {
     public class HttpRequestCountMiddleware : HttpRequestMiddlewareBase<Counter>
     {
         public HttpRequestCountMiddleware(RequestDelegate next, Counter counter)
-            : base(counter)            
+            : base(counter)
         {
             _next = next ?? throw new ArgumentNullException(nameof(next));
 
@@ -18,12 +19,9 @@ namespace Prometheus.HttpExporter.AspNetCore.HttpRequestCount
         {
             await _next(context);
 
-            var labels = GetLabelData(context);
-
-            if (labels != null)
-            {
-                _requestCount.WithLabels(labels).Inc();
-            }
+            _requestCount
+                .WithLabels(GetLabelData(context))
+                .Inc();
         }
 
         private readonly RequestDelegate _next;
