+++ Prometheus.HttpExporter.AspNetCore/Library/HttpExporterMiddlewareExtensions.cs
@@ -4,7 +4,7 @@ using Prometheus.HttpExporter.AspNetCore.HttpRequestCount;
 using Prometheus.HttpExporter.AspNetCore.HttpRequestDuration;
 using Prometheus.HttpExporter.AspNetCore.InFlight;
 
-namespace Prometheus.HttpExporter.AspNetCore
+namespace Prometheus.HttpExporter.AspNetCore.Library
 {
     public static class HttpExporterMiddlewareExtensions
     {
